Removed Bin Option
Instead of using binary files and then having to choose what arch
you're on the drone package was created.  Each drone will have their
own file to parse the XML data and then use the interface project to
build the project to get ready for Lair.  Once the lair project is
ready it is then imported in the base drone package.

Have to figure out why r.ParseMultipartForm(0) is having an issue with
bigger files.